requirements.txt
bcrypt==4.0.1
cffi==2.0.0
cryptography==46.0.3
duckdb
paramiko==3.3.1
psutil==5.9.6
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
